Refactor test data into own class. Tidy up.
