+++ gnd/src/androidTest/java/com/google/android/gnd/persistence/remote/FakeRemoteDataStore.java
@@ -1,14 +1,26 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.google.android.gnd.persistence.remote;
 
+import com.google.android.gnd.FakeData;
 import com.google.android.gnd.model.Mutation;
 import com.google.android.gnd.model.Project;
 import com.google.android.gnd.model.User;
 import com.google.android.gnd.model.feature.Feature;
-import com.google.android.gnd.model.form.Element;
-import com.google.android.gnd.model.form.Field;
-
-import com.google.android.gnd.model.form.Field.Type;
-import com.google.android.gnd.model.form.Form;
 import com.google.android.gnd.model.layer.Layer;
 import com.google.android.gnd.model.layer.Style;
 import com.google.android.gnd.model.observation.Observation;
@@ -24,37 +36,19 @@ import javax.inject.Inject;
 
 public class FakeRemoteDataStore implements RemoteDataStore {
 
-  private final Element textElement = Element.ofField(Field.newBuilder()
-    .setId("TEST_FIELD")
-    .setLabel("Test field")
-    .setRequired(true)
-    .setType(Type.TEXT)
-    .build());
-
-  private final Layer layerWithTextField = Layer.newBuilder()
-    .setId("LAYER_TEXT_FIELD")
-    .setName("Layer text field")
-    .setForm(Form.newBuilder()
-      .setElements(ImmutableList.of(textElement))
-      .build())
-    .setDefaultStyle(Style.builder()
-      .setColor("#0000ff")
-      .build())
-    .build();
-
   private final Layer layerWithNoFields = Layer.newBuilder()
-    .setId("LAYER_NO_FIELDS")
-    .setName("Layer no fields")
+    .setId(FakeData.LAYER_NO_FIELDS_ID)
+    .setName(FakeData.LAYER_NO_FIELDS_NAME)
     .setDefaultStyle(Style.builder()
-      .setColor("#00ff00")
+      .setColor(FakeData.LAYER_NO_FIELDS_COLOR)
       .build())
     .build();
 
   private final Project testProject = Project.newBuilder()
-    .setId("TEST_ID")
-    .setTitle("TEST TITLE")
-    .setDescription("TEST DESCRIPTION")
-    .putLayer("LAYER_NO_FIELDS", layerWithNoFields)
+    .setId(FakeData.PROJECT_ID)
+    .setTitle(FakeData.PROJECT_TITLE)
+    .setDescription(FakeData.PROJECT_DESCRIPTION)
+    .putLayer(FakeData.LAYER_NO_FIELDS_ID, layerWithNoFields)
     .build();
 
   @Inject
@@ -75,20 +69,6 @@ public class FakeRemoteDataStore implements RemoteDataStore {
   @Override
   public Flowable<RemoteDataEvent<Feature>> loadFeaturesOnceAndStreamChanges(
     Project project) {
-
-    /*Point point = Point.newBuilder()
-      .setLatitude(51.510357)
-      .setLongitude(-0.116773)
-      .build();
-
-    Feature feature = Feature.newBuilder()
-      .setId("FEATURE_ID")
-      .setProject(project)
-      .setPoint(point)
-      .setLastModified(AuditInfo.now(FakeAuthenticationManager.TEST_USER))
-      // Need to add layer info
-      .build();*/
-//    return Flowable.just(RemoteDataEvent.loaded("ENTITY_ID", feature));
     return Flowable.empty();
   }
